+++ internal/pkg/commands/cmd_runner_unix.go
@@ -47,6 +47,18 @@ func (cr *UnixCMDRunner) BuildDockerComposePullCMD(options DockerComposePullOpti
 	return Command{Cmd: command}
 }
 
+func (cr *UnixCMDRunner) BuildDockerComposeCreateCMD(options DockerComposeCreateOptions) Command {
+	services := strings.Join(options.Services, " ")
+	command := fmt.Sprintf("docker compose -f %s create %s", options.Path, services)
+	return Command{Cmd: command}
+}
+
+func (cr *UnixCMDRunner) BuildDockerComposeBuildCMD(options DockerComposeBuildOptions) Command {
+	services := strings.Join(options.Services, " ")
+	command := fmt.Sprintf("docker compose -f %s build %s", options.Path, services)
+	return Command{Cmd: command}
+}
+
 func (cr *UnixCMDRunner) BuildDockerPSCMD(options DockerPSOptions) Command {
 	command := "docker ps"
 	if options.All {